chore(*): move task requires into handler
This speeds up the cli warm up time from 3s to 0.2s.

Closes https://github.com/INSIDEM2M/slim/issues/39
